LayoutTests/media/modern-media-controls/volume-slider/volume-slider-value-expected.txt

Testing the VolumeSlider value property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS slider.value is 0

Now we'll drag to the slider thumb from 10px to 30px within the slider's bounds
controlValueWillStartChanging
controlValueDidChange, value = 0.019
controlValueDidChange, value = 0.024
controlValueDidChange, value = 0.03
controlValueDidChange, value = 0.035
controlValueDidChange, value = 0.041
controlValueDidChange, value = 0.046
controlValueDidChange, value = 0.051
controlValueDidChange, value = 0.057
controlValueDidChange, value = 0.062
controlValueDidChange, value = 0.068
controlValueDidChange, value = 0.073
controlValueDidChange, value = 0.078
controlValueDidChange, value = 0.084
controlValueDidChange, value = 0.089
controlValueDidChange, value = 0.095
controlValueDidChange, value = 0.1
controlValueDidChange, value = 0.105
controlValueDidChange, value = 0.111
controlValueDidChange, value = 0.116
controlValueDidChange, value = 0.122
controlValueDidStopChanging
PASS successfullyParsed is true

TEST COMPLETE
